requirements.txt
Django==3.2.23
djangorestframework
psycopg2
psycopg2-binary
django-filter
django_extensions
python-decouple 
django-cors-headers
gunicorn
arrow==1.3.0
chardet==5.2.0
contourpy==1.1.1
cycler==0.12.1
DateTime==5.4
fonttools==4.50.0
importlib-resources==6.3.1
kiwisolver==1.4.5
matplotlib==3.7.5
numpy==1.24.4
oauthlib==3.2.2
packaging==24.0
pandas==2.0.3
pillow==10.2.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
reportlab==4.1.0
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
types-python-dateutil==2.9.0.20240316
tzdata==2024.1
zipp==3.18.1
zope.interface==6.2
PyPDF2
python-decouple
-e ./python_withings_api